Support negative n-best, which samples from all hypotheses
Fixes #73.

+++ sentencepiece-sys/src/ffi/sentencepiece.h
@@ -16,7 +16,7 @@ int spp_decode_pieces(SentencePieceProcessor *spp, char const * const *pieces, s
 
 unsigned char *spp_encode_as_serialized_proto(SentencePieceProcessor *spp, char const *sentence, size_t sentence_len, size_t *len);
 
-unsigned char *spp_sample_encode_as_serialized_proto(SentencePieceProcessor *spp, char const *sentence, size_t sentence_len, size_t *len, size_t nbest, float alpha);
+unsigned char *spp_sample_encode_as_serialized_proto(SentencePieceProcessor *spp, char const *sentence, size_t sentence_len, size_t *len, int nbest, float alpha);
 
 SentencePieceProcessor *spp_new();
